An example on name mangling
This is a demonstration of how you can use an invoke C code in C++
code. While running this example make sure that you generate the object
codes for C and C++ separately, by using the "-c" option. Once object
codes are generated you can compile the codes with the "*.o" files.